pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "example-hermes-plugin"
version = "0.1.0"
description = "Example Hermes plugin — demonstrates tools, hooks, and bundled skill"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = []

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pyyaml>=6.0",
    "hermes-agent @ git+https://github.com/NousResearch/hermes-agent.git",
]

[project.entry-points."hermes_agent.plugins"]
example_plugin = "plugin"

[tool.setuptools.packages.find]
include = ["plugin*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
